move tests from `tests/js/plugins-and-defaults/versions.test-d.ts` into `tests/js/plugins-and-defaults/index.test-d.ts`
